revert radix to rounded floats
This pretty much defeats the point of using it, but it provides a point from which to
re-build a proper int version

+++ src/radixheap.cpp
@@ -46,7 +46,7 @@ RadixHeap::~RadixHeap()
     delete [] u;
 }
 
-void RadixHeap::insert(int item, long k)
+void RadixHeap::insert(int item, float k)
 {
     RadixHeapNode *newNode = new RadixHeapNode;
     newNode->item = item;
@@ -60,7 +60,7 @@ void RadixHeap::insert(int item, long k)
 #endif
 }
 
-void RadixHeap::decreaseKey(int item, long k)
+void RadixHeap::decreaseKey(int item, float k)
 {
     RadixHeapNode *node;
     node = nodes[item];
